+++ requirements.txt
@@ -1,1 +1,2 @@
 discord.py==1.7.1
+pytest~=6.2.3
